ChatBotPractice/requirements.txt
openai
pathlib
sounddevice
soundfile
numpy
pyyaml
gtts
google.generativeai